+++ frontend/src/api/v3/common.js
@@ -30,6 +30,12 @@ export const getReaderBasePath = (pk) => {
   return window.CODEX.API_V3_PATH + getReaderPath(pk);
 };
 
+export const getBookInBrowserURL = (pk) => {
+  const BASE_URL = window.CODEX.APP_PATH + getReaderPath(pk);
+  const timestamp = getTSParams().ts;
+  return `${BASE_URL}/book.pdf?ts=${timestamp}`;
+};
+
 export const getTSParams = () => {
   return { ts: useCommonStore().timestamp };
 };
@@ -41,6 +47,7 @@ const getVersions = (ts) => {
 
 export default {
   downloadIOSPWAFix,
+  getBookInBrowserURL,
   getReaderBasePath,
   getReaderPath,
   getTSParams,